Updated openstack/openstack Project: openstack/nova  f1d8bad73ba5758e5cf3387c406626216970daf8
Add missing foreign keys for sqlite

Previous migrations have added foreign keys for MySQL and PostgreSQL
but have missed SQLite.

This patch also fixes a unit test that was expecting results to be
returned from SQLite in a certain order. Adding a foreign key made
this non-deterministic.

Change-Id: I61baae56bb425400e917b1cb18815ca94280f4cb
Partially-implements: blueprint validate-migrations-and-model