Updated openstack/openstack Project: openstack/nova  c464511cc25f39f4b10a6e5911636adaad45ea34
Enforce DB model matches results of DB migrations

oslo.db offers a unit test that can be used to compare the results of
DB migrations against the model defined in Nova. Now that all of the
technical debt surrounding inconsistency between the model and the
migrations have been resolved, this check can be enforced.

Change-Id: I2f544add25cc0ae1d5b361c6757766271e160f45
Implements: blueprint validate-migrations-and-model