Updated openstack/openstack Project: openstack/nova  e9c826c59d1654d0a1cbb11907848f7ee4f1d9f6
Fix leaking exceptions from scheduler utils

scheduler_utils.setup_instance_group() can raise NoValidHost
exception and the conductor does not handle that properly which
causes the server to get stuck in a scheduling state instead of
going to ERROR.

The function is called several times in conductor/manager.py. This
patch moves the function calls in the try blocks already present.
It handles the missing Affinity filters as UnsupportedPolicyException,
which is a new exception type added here. This way the exception
raised by scheduler_utils.setup_instance_group() will cause the
VM to end up in ERROR state instead of NOSTATE, as it is expected.
Quota rollback is also handled properly, when it is needed.

Closes-bug: #1408326
Change-Id: I3a8ba37ded8ac3fa9b5f3f35b09feb82d822d583