Updated openstack/openstack Project: openstack/nova  089cea5b7c4ec7a6dd24dadc64858e4a9832ba30
object: serialize set to list

The NovaObjectSerializer is changed to serialize set iterables to
list instead of tuple. This was needed to ensure that every rpc
driver work in the same way. The normal rpc driver coverts tuples
to lists anyhow so this modification does not change the behavior
in production.
In the functional test environment the fake rpc driver does not
change anything on the message and it caused the filter scheduler
to receive a tuple in the filter_properties through rpc and
failed to append items to it.

This change is needed to use the functional test environment for
testing the filter scheduler.

Change-Id: I889badd85c96e0efadb3bc05f43c47bb995c9dce