Updated openstack/openstack Project: openstack/nova  386bc9376657a737e588640f41985ded07ee98c8
libvirt: avoid setting the memnodes where when it's not a supported option

The current version of Libvirt in the gate doesn't support a memnode element
in the numa tune configuration. When requesting guest topology the instance
is crashing. Not setting the numa tune memnodes option when libvirt version
is less than the required minimum.

Change-Id: I99d0bb64bd6d698f2d5ce978cb3778bbad8eec91
Closes-Bug: #1415333