Updated openstack/openstack Project: openstack/nova  7ee95ed19db6b63b12e164c4484dfed71a4bb0d6
Save bdm in swap_volume

swap_volume does not save bdm changes for connection_info. In
swap_volume, a call is made to _connect_volume that modifies
connection_info. For example, in the LibvirtISCSIVolumeDriver,
connection_info['data']['host_device'] = host_device. However,
this update is not saved to the DB, such that any updates made
within swap_volume are lost. Changes made in _connect_volume
need to be saved to the DB.

Change-Id: I19d5182d11097c4796093d77b5b7de11b7bc7bd8
Closes-Bug: #1375531