Updated openstack/openstack Project: openstack/nova  ebc8c63f1eb952e268a94bf2ac8c736a494d8faf
nova net-delete network is not informative enough

Intially nova net-delete <Network ID> was raising an
"InternalServerError" in case the requested network to
delete was associated to any project.This error category
“InternalServerError” is appropriate but the error
message description is not informative enough.

Now, when the command is not able to delete the requested
network because it is associated with a project, a message
stating “network must be disassociated from project
<project-id> before delete” will be raised. Updated the
code to raise the new error message.
Closes-Bug: #1301373

Change-Id: I82a6f8783ffce70934409177993de477b1d25da1