Updated openstack/openstack Project: openstack/neutron  c8b156fb77a66a6f0ec67884ee2e32436fcc4728
Encapsulate DVR Fip namespace

The floating ip namespace is an important concept that links a DVR
enabled L3 agent to DVR routers.  The namespace is shared by all of
the dvr routers on a compute host connected to the same network.  It
is managed by the dvr enabled agent.  Hence, it should be known to the
agent and made available to the DVR routers as well.  This calls for a
proper encapsulation of this functionality in a class which can be
accessed by both the agent and the routers.

In trying to tease floating ip functionality out of the L3 agent and
in to the new router classes, I found it very difficult to do because
the logic around the floating ip namespace has not yet been properly
encapsulated.

Change-Id: I0913cfae391164146bbb50ed42311506d6559e3f
Partially-Implements: bp/restructure-l3-agent