Updated openstack/openstack Project: openstack/neutron  88d6d61586ec5b8ef0334c83dd9d5f51676921c9
Don't pass the port down to the floating ip processing

This is justified by the Law of Demeter.  I ran in to this while
trying to refactor the floating ip processing here [1].  The scope of
that patch can be reduced significantly with this refactor.  Also, it
helps us to decouple the floating ip NAT processing from the details
of the port which is a long-term goal.

[1] https://review.openstack.org/#/c/142863/

Change-Id: I606dfd43977934f1accea60911e10e71c1cbba73
Partially-Implements: bp/restructure-l3-agent