Updated openstack/openstack Project: openstack/nova  dc0436194d44bec6980753e5fea5aa499709eef7
initialize objects with context in block device

These changes aim to clean up the pattern of passing a context in
object member functions create/destroy/refresh/save and instead
initialize the object with the context when it's constructed.

Related to blueprint kilo-objects

(cherry picked from commit 570c98cb9e9e7327477d92c52bfdc0a28ab7f88e)

This patch was reverted in https://review.openstack.org/#/c/147751/
because it broke boot from block device functionality, the test for
which is currently being skipped in tempest, temporarily. I tested
this locally with tempest with test_volume_boot_pattern enabled.

This is a resubmission of the reverted patch with the broken call
site fixed.

This clean up is intended to reduce complexity and confusion about
the use of objects, and in the future, objects will no longer
give the option of passing a context to create/destroy/refresh/save.
Instead, the context manager obj_as_admin() should be used to perform
an action with an elevated context.

Change-Id: I598df35d055b39f25b28e7bcff1062b875788ded