Updated openstack/openstack Project: openstack/tempest  3fc44f236325418868ccc8dc06df6f9d240e938f
Add a tests for service clients

Now we are trying to migrate service clients to tempest-lib,
and we need to check all arguments of each service client work
fine as library functions.
This patch adds a test for doing it.

There is data_utils module in Tempest for generating random
data, but this test doesn't use it because this test will be
moved to tempest-lib in the future and we are not sure whether
data_utils also is moved or not.

Note:
  If adding migrated client to test_clients which is added by
  this patch on each migration patch, we can test each client
  migration after this patch.

Change-Id: Ic6e1efd6ff9931e16a4ba490ef4aba4a23f71370