Updated openstack/openstack Project: openstack/neutron  e7e5b57bf806b7ec5b4450329d09d67cd44b5ece
NSX DB models split, part 1

This patch introduces a new module for hosting NSX DB model classes,
and populates it with models for the 'network gateway' extension,
which are then removed from the model hosting the network gateway
mixin.

This patch is part of a set of minor refactorings in the NSX plugin
which will then lead to a semi-full decomposition by the end of
the Kilo release cycle.

Related-to blueprint core-vendor-decomposition

Change-Id: Ie04f6c94055ac20b938c17cea251ec6532a1e16a