Updated openstack/openstack Project: openstack/neutron  c2aa9108ccadd6cb9c17e0e9b28769f8fbfde69d
Make prevent_l3_port_deletion handle missing port

This adjusts the prevent_l3_port_deletion function to handle
the case where the port ID that is passed to it does not have
an entry in the database.

Previously it was raising an exception in this case, which is
inconsistent to how ML2 was handling concurrent port_delete requests
further in the port delete function (log them but don't fail).

Closes-Bug: #1416554
Change-Id: I6da021bdf0c79f72336416d02ab989407f352904