Updated openstack/openstack Project: openstack/tempest  11fbf0f3ecb3cf4f1eecfb08b4818c739c2368df
Remove data_utils import from a service client

There is data_utils module in Tempest for generating random
data, but it is better to avoid using it in service clients
because service clients will be moved to tempest-lib in the
future but we are not sure whether data_utils also is moved
or not. Then, this patch removes data_utils import from the
messaging client.

Change-Id: I2070d33ea8d174092981dca3e59fe65d13a13fa7