Updated openstack/openstack Project: openstack/neutron  014b4be5688479661c5b185224a1d279f95e4b94
Refactor ml2 manager

Refactor code to facilitate other changes. Keeping the
refactored code in different patch for ease of review.
This patch mainly consolidates the keys that are referenced
together most of the time in reference to providernet
extension. By doing this, the subsequent changes are less
repetitive and makes the code little cleaner as well.

Change-Id: Idc0648d5c4688c8f797cc5427b71c2a3919ce722
Related-Bug: #1333475