Updated openstack/openstack Project: openstack/nova  bb1a6b6dda291d40f5a02723b7ef46c4ac8d747c
libvirt : Add support for --interface option in iscsiadm.

Adds the new libvirt parameter iscsi_transport that can be used to
specify an iscsi transport, which used in conjuction with the
--interface parameter provides offloaded iscsi support.

Also happens to implement code that was originally supposed to be
covered by hw-iscsi-device-name-support as this is a
requirement for transport support.

DocImpact
Closes-Bug: #1370226
Implements: blueprint hw-iscsi-device-name-support
Implements: blueprint add-open-iscsi-transport-support

Change-Id: I1034f1e26e0b00e64430e6347d232793c3401ba8