Updated openstack/openstack Project: openstack/cinder  e6519f227f8c1b6d7ef6f0863a762c8faea75b89
Roll back if VMAX masking view not created

In emc_vmax_masking.py, an exception object is created if
there is a failure in the masking view creation logic.
However, the exception is not raised. As a result,
initialize_connection will succeed in spite of a failure
but the VM will not be able to access the disk.
This patch adds rollback logic to handle failures in
creating masking view. If the masking view cannot be
created, the volume will be added back to the default
storage group and an exception will be raised.

Closes-Bug: #1393540
Change-Id: I1a00d8922cc876a943fb7822cd86c580d8e6beac