Updated openstack/openstack Project: openstack/horizon  4ae40a803eb8f91f44ed57e2dad8133b064eedbc
View encryption metadata for encrypted volumes

Although it is possible to see whether a volume is encrypted or not
encrypted (by viewing the "Encrypted" column of the volume), it is
not yet possible to view the encryption metadata for an encrypted
volume.

This patch adds a link to the "Yes" in the Encrypted column of the
volume, which displays an overview of the encryption metadata for
the volume when clicked.  This encryption metadata includes cipher,
key size, provider, control location, and volume type name.  Also
adds unit tests.

Change-Id: Ic57e64eb85b46e77ed03d86722523c316c873d18
Implements: blueprint view-volume-encryption-metadata