Updated openstack/openstack Project: openstack/keystone  a8686408681ecd8b13fe247989dfc75c8ef9bb58
Refactor keystone-all and http/keystone

There was a lot of code in keystone-all and httpd/keystone.
Functions in keystone-all and http/keystone can't be called from the
tests, so this code should be in the keystone package.

This also allows some sharing of common code between keystone-all,
httpd/keystone, and the tests.

bp refactor-keystone-all-http

Change-Id: I1c4e59e253b1816ccfb4d5bf1d2aa40b49221b4f