Updated openstack/openstack Project: openstack/swift  49586c6ed7369f0195292d4cc827a8c305fea38e
Prevent redundant commenting by drive-audit

The drive-audit detects error log about a device and comments out it
in /etc/fstab. When the error log is generated several times, it
comments out the line for each time.
This patch makes drive-audit to check if the device is already
commented out, and prevents redundant commenting out.

Change-Id: Ia542d35b58552dde0f324bb9c42531f98c9058fa