Updated openstack/openstack Project: openstack/nova  a219393c8c6cb935721e45271f8108c0a51571db
Don't create block device mappings in the API cell

Otherwise 2 block_device_mapping entries will be created
in the API cell:
- the first one (created by the API cell) will have close to
  no information about the volume (device_name and volume_id are NULL)
- the second one (bubbled up from the compute cell) will contain
  all the volume information

The first entry confuses Nova when creating an image since
it won't be able to find the associated volume (NULL) in Cinder.

The compute cell should create it first and propagate it up
to the API cell.

Change-Id: I38edb953e73de6bc70a2e5950c68f457f83303e1
Closes-bug: #1417239