Updated openstack/openstack Project: openstack/nova  26b8ac554902a5a1e4310641a4386518163a8fd5
Cells: Fix service_get_by_compute_host

Compute_node objects require that the 'id' field be an int, but with
cells the 'id' is a string with routing information.  So conversion of
the dict returned via RPC from a child cell fails.

This patch adds a proxy object to keep the routing information visible
via the API without changing the underlying object model.

Change-Id: I71ab82841983798105237804d20bd41c04acfc7b
Closes-bug: #1408122