Updated openstack/openstack Project: openstack/python-neutronclient  7aaf4a48ab59687ae6c8b9cddfb4f05f4a41aa9a
Skip None id when getting security_group_ids

When getting security_group_ids in securitygroup.py, a None id is
included into the list due to the remote group id is Null when the
security group rule direction is egress.
And this causes the calculation of the number of chunk_size is
1 more than the number it should be.
Then the request page size excess the MAX request page size
Then the RequestURITooLong exception happens again, and finally
the commend fails.

Co-Authored-By: Furukawa, Yushiro <y.furukawa_2@jp.fujitsu.com>
Change-Id: Ib671cf3ddc6827793c4a86d2d3c8a34957f5dfd3
Related-Bug: 1271462
Closes-Bug: 1415692