Updated openstack/openstack Project: openstack/nova  108cf911a34820345673c10993ac34488d5c4252
Replace select-for-update in fixed_ip_associate_pool

MySQL Galera does not support the write-intent locks
that SELECT FOR UPDATE requires. In this patch SELECT
FOR UPDATE in 'fixed_ip_associate_pool' function
was replaced with SELECT and UPDATE.

The problems with MySQL Galera and algorithm of replacing
SELECT FOR UPDATE relate to spec
https://review.openstack.org/#/c/135296

Also this patch fixes incorrect network assigning
because previously the relationship 'network' was
updated instead of column 'network_id'.

Change-Id: I403da67a8d55aa30c9abcc491f71918cd2b452dd