Updated openstack/openstack Project: openstack/python-neutronclient  b510624700e0c589a53337797d9b0ce7924df833
Utility method for boolean argument

In the recent commit, True/False of boolean argument
became case-insensitve and this code is copy-and-paste'ed
in several places.

Now there are several number of patches which add explicit
arguments for *-update or *-list and the code of boolean
argumetns like below will be used more than now.

This commit add a utility method to register such boolean opt.

       parser.add_argument(
           '--enabled',
           dest='enabled', metavar='{True,False}',
           choices=['True', 'true', 'False', 'false'],
           help=_('Whether to enable or disable this rule.'),
           default=argparse.SUPPRESS)

Change-Id: I9575eeef32154a8b92589c2cc7889803216bddb2