Updated openstack/openstack Project: openstack/neutron  5132303bbea3f9cf4392eb68b6082c72484fee92
Fix extra-dhcp-opt on stateless dhcpv6 subnet

The opts in dnsmasq opt-file is depending on the tag set
in the host-file. When you have only one stateless dhcpv6 subnet,
it will be filtered by the _iter_host() function, and will not be
written to the host-file, this may lead to extra-dhcp-opt not working.

Closes-Bug:#1411137

Change-Id: Ieed16280fb3c4c91f7d906c0410446d2a112b93e