Updated openstack/openstack Project: openstack/nova  7e0e84e1a102381221d8c22af28498d49ac342be
Fix corrupting the object repository with test instance objects

In order to test older instance versions in the remote instance
tests, we were creating an OldInstance object masquerading as a
proper Instance so we could override the version. That had the
potential to confuse other instance object tests because from that
point on, it was the first Instance object in the repository,
breaking the ordering requirement. Since we don't need it to be
called Instance to test what we want to test, this patch just
removes the masquerading bit.

Change-Id: Iabb77ea3084c3701bf67f051ace494a1efb12917
Closes-Bug: #1417678