Updated openstack/openstack Project: openstack-dev/devstack  da352c7b4bbe8fb7f8f663f84194f7cddbe8f3e2
Ceph: wait for services to shutdown during cleanup

It can take ceph-osd and ceph-mon a few seconds to complete when
they are killed.  This races against the umount command in cleaup
and can often result in $CEPH_DATA_DIR failing to unmount since
it is still in use.

Wait for these processes to stop to ensure the mount point is
umounted successfully.

Change-Id: I1a635e75a68be6b14fbee52ff981b5f5a3a8eb0e