Updated openstack/openstack Project: openstack/nova  f88a13684d800cb2703c1d2ec6e9452aa0313c0f
Do not treat empty key_name as None

When running instance, empty ("") key_name is treated as None and
keypair lookup does not happen.  However, empty string is written to
key_name field in instance's database record.  When horizon renders
instance info, it looks up keypair if key_name is not None.  Having
empty string in this property generates error and instance info is not
displayed.

Change-Id: Ib67b124e6480f3435d31c983f1bfca769899c8e7
Closes-Bug: #1403544