Updated openstack/openstack Project: openstack/horizon  52d5673e363b7204ec55f292863bbbe43ce0cf3c
added detail page for keypair

from keypair table, we can see only the
name and fingerprint of the keypair

in horizon we have no option to see the
id, created time and public key of
the keypair.

this patch adding a new detail page for
keypairs, which display the name fingerprint,
id, created time and public key.

Change-Id: Idb85f7e60090875e2e503806122cdb4ff637c341
Closes-Bug: #1413126