Updated openstack/openstack Project: openstack/nova  65aef78b8469ca14841ccd15c1024ab3cff122fb
Add formal doc recording hypervisor feature capability matrix

Add document to replace / obsolete the giant table on

  https://wiki.openstack.org/wiki/HypervisorSupportMatrix

This initial draft is a fairly straightforward conversion of
that table. Over time, it needs much work to improve the coverage
of API operations and and coverage of important configuration
information that users will care about.

It is using the .ini file syntax in order to record the data in
an easily machine parsable format, while remaining human friendly
by avoiding the syntax heavy approach of XML / JSON / YAML

An extension is registered with sphinx that can convert the
.ini file content into docutils content that then gets rendered
into the developer docs, linked from the index page

Change-Id: I4d3db4bce5737dba30a026a11083a9ea64459cd4