Updated openstack/openstack Project: openstack/glance  d55522da7373e8e046f63ebc54a22557c3dd6267
Software Metadata Definitions

The will provide a base library of metadata definitions for
various common software products, components, and libraries that
may exist on particular image (or volume or instance). These
metadata definitions will make it easier for end users and
admins to easily describe the software and its properties.
This information will enable an improved and faster user
experience for applying software metadata, searching based
on software metadata, and viewing the software information
about an image.

Various improvements in horizon are underway to take
advantage of this metadata. For example, a user launching
an instance will be able to expand the image row to see
additional information about the image.  This will include
providing the metadata definition for properties on the image.
This same information will also be visible from the image and
instance details page.

Change-Id: I63fa40f9eae8dfb602cf0c9c3f0dd7f0d1a65af0
Implements: blueprint software-metadefs