Updated openstack/openstack Project: openstack/django_openstack_auth  e6c25ad380dc1feb0ed121d088151d9fde8cacef
Create plugin model for DOA authentication

With federated and kerberos logins coming we need an extensible way to
specify additional ways to fetch an unscoped token from keystone.

Create a plugin model that when authenticate is called a series of
plugins can be queried for a token depending on the information
provided.

Closes-Bug: #1433389
Change-Id: Ifbd7077173844a8eb3400799fd512b62a5dc7dcc