Updated openstack/openstack Project: openstack/python-cinderclient  9dcb38f5f2fd3712f97e29c6c460773376d0ac73
Fix up help message for reset-state call

There's some confusion about what the reset-state
call does.  It's not clear to some that this call
simply changes the state of the object in the data-base.

This patch just updates the help message to indicate more
clearly that reset-state ONLY changes state in the DB.

Change-Id: Ia389cfd8b1a821a356706201c009cf3c8dca8419