Updated openstack/openstack Project: openstack/nova  696ae837a425ee7a619d41a29ff7d12faec39158
libvirt: add ploop disks format support

The way how PCS works with disks slightly differs from
how qemu does. In PCS each disk has metadata, stored in
a separate xml file. The most important thing in this
metadata is paths and types of base image and deltas.
The name of xml file is fixed (DiskDescriptor.xml) and
usually all images are stored in the same directory.
Images are being downloaded from glance to cache in a form
of files. So imagecache will work in the same way as with Raw
image type.

Here is and example of libvirt's disk config:

    <disk type='file' device='disk'>
      <driver type='ploop'/>
      <source file='/path/to/dir/with/xml'/>
      <target dev='sda' bus='sata'/>
    </disk>

Since you need to provide a directory with xml file to
libvirt, separate class is needed imagebackend.py for
working with such images.

This patch introduces class Ploop which are similar to Raw,
it copies base image to the instance's dir and prepares it
for starting VM or container from this image.

If force_raw_images is false, then only raw and ploop image
types allowed. If force_raw_images is true, then all image
types, which are supported by qemu-img can be used.

Partially implements blueprint pcs-support

Change-Id: I9c2a0da52dbda7d45f749f9d42b2760a0d2e790f