Updated openstack/openstack Project: openstack/nova  0c3d88add0d7eafb0c936f13f9ad79113716d7be
Add missing setup.cfg entry for os-user-data plugin

Although os-user-data has an entry in nova.api.v3.extensions.server.create
in setup.cfg it is missing one in the general nova.api.v3.extensions list.
Without an entry the plugin does not appear in the extension list so this
patch adds an entry for it.

Partially implements blueprint v2-on-v3-api

Change-Id: I13fa049539b07ff901b36ae18eca59fbb1354794