Updated openstack/openstack Project: openstack/nova  549881ec2d5e39d798d5564f3b7841df6e0f4d0d
Add max_concurrent_builds limit configuration

Right now, nova-compute will attempt to build an infinite number of
instances, if asked to do so. This won't work on any machine, regardless
of the resources, if the number of instances is too large.

We could default this to zero to retain the current behavior, but
the current behavior is really not sane in any case, so I think we
should default to something. Ten instances for a single compute node
seems like as reasonable default. If you can do more than ten at a
time, you're definitely not running a cloud based on default config.

DocImpact: Adds a new configuration variable

Closes-Bug: #1418155

Change-Id: I412d2849fd16430e6926fc983c031babb7ad04d0