Updated openstack/openstack Project: openstack/nova  f47676c21ebbd18d708ed547f54a10128d53ed29
libvirt: fix console device for system z for log file

The system z platform needs a console device for the log file of an
instance. A (manual) regression test showed that this is not the case
with master from 2015-01-22. The guest.cpu.arch is 'None' which leads
to the creation of a wrong device. The bug was introduced with commit
11a2ed096

This fix uses the libvirt utils to decide which architecture has to be
considered.

Closes-bug: 1414038

Change-Id: I97dc55e6fca0b2491d4a9bc81366f9c92d39002a