Updated openstack/openstack Project: openstack/nova  98641da6402c44793da4d2f6d7ea1a7ba5f16708
Ironic: Adds config drive support

This patch is adding config drive v2 support for the Ironic drive. As
attaching a volume in a baremetal instance is not as simple as VMs and
also hardware depedent, this patch will base64 + gzip the generated
configdrive and pass it to Ironic as a string, Ironic then can take
care about storing and exposing it to the the tenant of baremetal.

DocImpact
Implements: blueprint use-configdrive-with-ironic
Change-Id: Ibc745e9a63bc3d5d68217020b3d591f71c3f8a96